Revert "Move `ContactgroupName` to store"
This reverts commit 08ac91be95b7991377eb098ea00cee1454aa23fd.
