+++ cmk/base/config.py
@@ -90,15 +90,11 @@ from cmk.utils.rulesets.ruleset_matcher import (
 )
 from cmk.utils.sectionname import SectionName
 from cmk.utils.site import omd_site
-from cmk.utils.store.host_storage import (
-    apply_hosts_file_to_object,
-    ContactgroupName,
-    get_host_storage_loaders,
-)
+from cmk.utils.store.host_storage import apply_hosts_file_to_object, get_host_storage_loaders
 from cmk.utils.structured_data import RawIntervalsFromConfig
 from cmk.utils.tags import ComputedDataSources, TagGroupID, TagID
 from cmk.utils.timeperiod import TimeperiodName
-from cmk.utils.type_defs import CheckPluginNameStr, Item, ServiceName
+from cmk.utils.type_defs import CheckPluginNameStr, ContactgroupName, Item, ServiceName
 
 from cmk.automations.results import CheckPreviewEntry
