+++ cmk/base/core_nagios.py
@@ -29,9 +29,8 @@ from cmk.utils.labels import Labels
 from cmk.utils.licensing.handler import LicensingHandler
 from cmk.utils.log import console
 from cmk.utils.macros import replace_macros_in_str
-from cmk.utils.store.host_storage import ContactgroupName
 from cmk.utils.timeperiod import TimeperiodName
-from cmk.utils.type_defs import CheckPluginNameStr, ServiceName
+from cmk.utils.type_defs import CheckPluginNameStr, ContactgroupName, ServiceName
 
 from cmk.checkengine import plugin_contexts
 from cmk.checkengine.check_table import FilterMode
