+++ cmk/base/default_config/base.py
@@ -10,11 +10,11 @@ from cmk.utils.labels import Labels
 from cmk.utils.notify_types import Contact, ContactName
 from cmk.utils.password_store import Password
 from cmk.utils.rulesets.ruleset_matcher import RuleSpec, TagsOfHosts
-from cmk.utils.store.host_storage import ContactgroupName, FolderAttributesForBase
+from cmk.utils.store.host_storage import FolderAttributesForBase
 from cmk.utils.tags import TagConfigSpec
 from cmk.utils.timeperiod import TimeperiodSpecs
 from cmk.utils.translations import TranslationOptionsSpec
-from cmk.utils.type_defs import CheckPluginNameStr, ServiceName
+from cmk.utils.type_defs import CheckPluginNameStr, ContactgroupName, ServiceName
 
 from cmk.snmplib.type_defs import SNMPCredentials
