+++ cmk/gui/userdb/store.py
@@ -30,7 +30,7 @@ from cmk.utils.store import (
     save_text_to_file,
     save_to_mk_file,
 )
-from cmk.utils.store.host_storage import ContactgroupName
+from cmk.utils.type_defs import ContactgroupName
 from cmk.utils.user import UserId
 
 import cmk.gui.hooks as hooks
