+++ cmk/utils/store/host_storage.py
@@ -18,12 +18,11 @@ from cmk.utils.hostaddress import HostName
 from cmk.utils.labels import Labels
 from cmk.utils.rulesets.tuple_rulesets import ALL_HOSTS, ALL_SERVICES
 from cmk.utils.tags import TagGroupID, TagID
+from cmk.utils.type_defs import ContactgroupName
 
 HostAttributeMapping = tuple[
     str, str, dict[str, Any], str
 ]  # host attr, cmk.base var name, value, title
-
-ContactgroupName = str
 
 
 class GroupRuleType(TypedDict):
